code/requirements.txt
imutils==0.5.4
numpy==1.21.2
opencv_python_headless==4.5.5.64
Pillow==9.2.0
torch==1.10.0+cu113
torchvision==0.11.1+cu113
